add first version of github actions workflows
Added two workflows: one for building and one for deploying.
1. Building is triggered whenever there are new commits on the master branch.
Since this repo has a protected master branch, this will only be done by merging
a PR onto it. The NPM version is bumped, new Docker images are built and a new
release is published.
2. Deploying is triggered whenever there is a new release in the repo published.
Then, the release's tag version is used to deploying said version on the K8s
cluster(s).