GitHub actions (#46)
* add first version of github actions workflows

Added two workflows: one for building and one for deploying.
1. Building is triggered whenever there are new commits on the master branch.
Since this repo has a protected master branch, this will only be done by merging
a PR onto it. The NPM version is bumped, new Docker images are built and a new
release is published.
2. Deploying is triggered whenever there is a new release in the repo published.
Then, the release's tag version is used to deploying said version on the K8s
cluster(s).

* make deployment files compatible w/ Kustomize v2

Files were only for Kustomize v3. For the new GitHub Actions we use, for
simplicity, the default kubectl tool, which only includes Kustomize v2. For this
reason, adapt the deployment files to support Kustomize v2.

* add script for CI

* refer to default-branch

* split deploy in seperate job per env

* start build and deploy manually instead of on each PR